requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
Brotli==1.0.9
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cssselect2==0.7.0
Deprecated==1.2.14
distro==1.8.0
Django==4.2
django-celery-beat==2.5.0
django-cleanup==8.0.0
django-phonenumber-field==7.2.0
django-ranged-response==0.2.0
django-robots==6.1
django-simple-captcha==0.6.0
django-simple-math-captcha==2.0.1
django-timezone-field==6.0
fonttools==4.42.0
html5lib==1.1
idna==3.4
kombu==5.3.1
netaddr==0.8.0
phonenumberslite==8.13.26
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.3
pycparser==2.21
pydyf==0.7.0
pyphen==0.14.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
wrapt==1.15.0
yookassa==2.4.0
zopfli==0.2.2
